Added check to see if backup successful

+++ src/backup/jackett.bkp.fish
@@ -20,12 +20,19 @@ end
 
 echo "jackett.bkp.fish -- Creating archive"
 tar -cvzf $arch -C $src/.. jackett
+if test $status -eq 0
+    logger -t jackett.rec.fish "The backup was successful"
+    echo "jackett.rec.fish -- The backup was successful"
 
-set nb_backups (command ls -1trd $dst/jackett.*.tgz | wc -l)
-set nb_backups_todelete (math $nb_backups - $nb_max_backups)
-if test $nb_backups_todelete -gt 0
-    echo "jackett.bkp.fish -- Removing older archives"
-    command ls -1trd $dst/jackett.*.tgz \
-        | head -n$nb_backups_todelete \
-        | xargs rm -f
+    set nb_backups (command ls -1trd $dst/jackett.*.tgz | wc -l)
+    set nb_backups_todelete (math $nb_backups - $nb_max_backups)
+    if test $nb_backups_todelete -gt 0
+        echo "jackett.bkp.fish -- Removing older archives"
+        command ls -1trd $dst/jackett.*.tgz \
+            | head -n$nb_backups_todelete \
+            | xargs rm -f
+    end
+else
+    logger -t jackett.rec.fish "Backup unsuccessful"
+    echo "jackett.rec.fish -- Backup unsuccessful"
 end
